+++ src/main/java/com/phoneshope/java/project/service/BrandService.java
@@ -10,9 +10,9 @@ import java.util.Map;
 public interface BrandService {
     Brand create(Brand brand);
 
-    Brand getById( Integer id);
+    Brand getById( Long id);
 
-    Brand update(Integer id,Brand brandUpdate);
+    Brand update(Long id,Brand brandUpdate);
 
 
     List<Brand> getAllBrands();
